Refactor CompetitionsList page
