+++ app/src/pages/CompetitionsList/containers/index.js
@@ -1,0 +1,3 @@
+export { default as Header } from './Header';
+export { default as Controls } from './Controls';
+export { default as List } from './List';
