+++ app/src/pages/CompetitionsList/context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const CompetitionsListContext = createContext({});
+
+export { CompetitionsListContext };
